did some mall refactoring
