+++ react/week4/github-users-search-app/src/components/SearchInput.js
@@ -1,5 +1,5 @@
 import React, { useContext } from "react";
-import { userContext } from "./userContext";
+import { userContext } from "../contexts/userContext";
 import DisplayUsers from "./DisplayUsers";
 
 export default function SearchInput() {
