+++ react/week4/github-users-search-app/src/components/searchUsersApi.js
@@ -5,4 +5,4 @@ const searchUsersApi = async (searchQuery) => {
   return API;
 };
 
-export { searchUsersApi };
+export default searchUsersApi;
